perf: upgrade to Next/React canaries and adopt Cache Components
Bump next to 16.3.0-canary and React to 19.3 canary, migrate daily puzzle,
blog, and leaderboard data to "use cache"/cacheLife/cacheTag, seed auth from
the server to remove the session waterfall, and serverize leaderboard and
game-over shells with loading states and idle guest prefetch.
