+++ apps/web/package.json
@@ -3,8 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --webpack",
-    "build": "next build --webpack",
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "build:webpack": "next build --webpack",
     "start": "next start",
     "lint": "npx biome check .",
     "lint:fix": "npx biome check --write .",
@@ -78,11 +79,11 @@
     "jose": "^6.1.2",
     "lucide-react": "^0.468.0",
     "mongodb": "^6.0.0",
-    "next": "^16.0.0",
+    "next": "16.3.0-canary.104",
     "next-themes": "^0.4.4",
-    "react": "^19.0.0",
+    "react": "19.3.0-canary-0f42eac2-20260730",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^19.0.0",
+    "react-dom": "19.3.0-canary-0f42eac2-20260730",
     "react-hook-form": "^7.54.1",
     "react-markdown": "^10.1.0",
     "react-resizable-panels": "^3.0.6",
@@ -104,8 +105,8 @@
     "@types/canvas-confetti": "^1.9.0",
     "@types/jest": "^30.0.0",
     "@types/node": "^20",
-    "@types/react": "^19.0.0",
-    "@types/react-dom": "^19.0.0",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "@types/uuid": "^11.0.0",
     "autoprefixer": "10.4.20",
     "babel-plugin-react-compiler": "^1.0.0",
